# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenISMS API",
    "description": "REST API für das Information Security Management System (OpenISMS).\n\n## Authentifizierung\nAlle geschützten Endpunkte erfordern einen **Bearer JWT-Token** im `Authorization`-Header:\n```\nAuthorization: Bearer <token>\n```\nDen Token erhalten Sie über `POST /api/auth/login`.\n\n## Rollen\n- `admin` – Vollzugriff\n- `assessor` – Risikobewerter / Auditor\n- `it-staff` – IT-Mitarbeiter (Security-Daten)\n- `dpo` – Datenschutzbeauftragter\n- `owner` – Asset-Verantwortlicher\n- `management` – Nur-Lesen (Management-Ebene)\n- `employee` – Schulungs-Tab, eigene Trainingsübersicht\n- `viewer` – Nur Lesezugriff\n\n**Custom Roles**: Eigene Rollennamen mit Basisrolle über `/api/admin/custom-roles` verwalten.",
    "version": "2.2.0",
    "contact": { "name": "ISMS Support" }
  },
  "servers": [
    { "url": "/api", "description": "Aktuelle Instanz" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Incident": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "ref": { "type": "string", "example": "INC-0001" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "category": { "type": "string", "enum": ["malware","phishing","data_breach","dos","unauthorized_access","misconfiguration","loss_theft","social_engineering","other"] },
          "severity": { "type": "string", "enum": ["low","medium","high","critical"] },
          "status": { "type": "string", "enum": ["reported","investigating","contained","resolved","closed"] },
          "nis2_reportable": { "type": "boolean" },
          "detected_at": { "type": "string", "format": "date-time" },
          "occurred_at": { "type": "string", "format": "date-time" },
          "resolved_at": { "type": "string", "format": "date-time", "nullable": true },
          "assignee_id": { "type": "integer" },
          "reporter_id": { "type": "integer" },
          "lessons_learned": { "type": "string", "nullable": true },
          "external_reference": { "type": "string", "nullable": true }
        }
      },
...
      "VvtEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "purpose": { "type": "string" },
          "legal_basis": { "type": "string" },
          "data_categories": { "type": "array", "items": { "type": "string" } },
          "recipients": { "type": "string", "nullable": true },
          "third_country_transfer": { "type": "boolean" },
          "dsfa_required": { "type": "boolean" },
          "status": { "type": "string", "enum": ["draft","active","archived"] },
          "last_reviewed_at": { "type": "string", "format": "date", "nullable": true }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/incidents": {
      "get": {
        "tags": ["Vorfälle"],
        "summary": "Vorfälle abrufen",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "severity", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Vorfallsliste", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Incident" } } } } } }
      },
      "post": {
        "tags": ["Vorfälle"],
        "summary": "Vorfall melden",
        "description": "Erfordert Rolle: admin, assessor, it-staff",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Incident" } } } },
        "responses": { "201": { "description": "Vorfall angelegt" } }
      }
    },
...
    "/incidents/{id}": {
      "get": { "tags": ["Vorfälle"], "summary": "Vorfall-Details", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Vorfalldetails" } } },
      "put": { "tags": ["Vorfälle"], "summary": "Vorfall aktualisieren", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Incident" } } } }, "responses": { "200": { "description": "Aktualisiert" } } },
      "delete": { "tags": ["Vorfälle"], "summary": "Vorfall löschen", "description": "Erfordert Rolle: admin", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Gelöscht" } } }
    },
...
    "/vvt": {
      "get": {
        "tags": ["DSGVO"],
        "summary": "Verarbeitungsverzeichnis abrufen (Art. 30)",
        "description": "Erfordert Modul: dsgvo",
        "responses": { "200": { "description": "VVT-Einträge", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/VvtEntry" } } } } } }
      },
      "post": {
        "tags": ["DSGVO"],
        "summary": "VVT-Eintrag anlegen",
        "description": "Erfordert Rolle: admin, assessor, dpo — Modul: dsgvo",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VvtEntry" } } } },
        "responses": { "201": { "description": "VVT-Eintrag erstellt" } }
      }
    },
    "/vvt/{id}": {
      "get": { "tags": ["DSGVO"], "summary": "VVT-Eintrag abrufen", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "VVT-Eintrag" } } },
      "put": { "tags": ["DSGVO"], "summary": "VVT-Eintrag aktualisieren", "description": "Erfordert Rolle: admin, assessor, dpo", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VvtEntry" } } } }, "responses": { "200": { "description": "Aktualisiert" } } },
      "delete": { "tags": ["DSGVO"], "summary": "VVT-Eintrag löschen", "description": "Erfordert Rolle: admin, dpo", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Gelöscht" } } }
    },
...
  },
...
}